refactor: move _permute_flat_segments to generation_multimodal
The helper exists only for flattened multimodal layouts; the text-core
generation_decoderonly module was an odd home for it. Its test moves to
the flatten suite alongside.
